+++ src/main/java/com/projectcourse/projectcourse/repository/LectureRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.projectcourse.projectcourse.entity.Lecture;
 
-public interface LectureRepository extends JpaRepository<Lecture , Integer> {
+public interface LectureRepository extends JpaRepository<Lecture , Long> {
     
 }
